requirements.txt
# LearnNova - Requirements
# AI-Powered Learning & Career Guidance Platform

# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0

# Database
SQLAlchemy>=2.0.35

# Celery for background tasks
celery==5.3.4
redis==5.0.1

# API & Serialization
marshmallow==3.20.1
flask-marshmallow==0.15.0
pydantic>=2.0.0
email-validator>=2.0.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Email
Flask-Mail==0.10.0

# Password hashing
Werkzeug==3.0.1

# File uploads
Pillow>=10.0.0

# Payment integration
razorpay==1.4.1

# AI/ML
openai==1.3.5
google-generativeai==0.3.2
scikit-learn>=1.3.2
numpy>=1.26.2
pandas>=2.1.4

# Transformer-based NLP
transformers>=4.36.0
torch>=2.1.0
spacy>=3.7.2
protobuf>=3.20.0,<4.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2

# Production
gunicorn==21.2.0
